Simplify the declaration-emit check to lean on tsc
The hand-rolled check script re-implemented the compiler's own error
reporting. `TS7056`/`TS4020` are ordinary declaration-emit errors, so a bare
`tsc -p tsconfig.declcheck.json` already fails the build on overflow — no
script needed for that half.

Reduce `check:declarations` to `rm -rf .tsdeclcheck && tsc -p
tsconfig.declcheck.json && node scripts/checkDeclarationSize.mjs`, leaving only
a small size canary that guards against a public type expanding inline before
it grows large enough to trip TS7056 on this small example backend.
